Rename scripts, remove unused ones
