--- scripts/distributions/archi/calculateLPDistributionsFromStakeFor.ts
@@ -1,558 +1,0 @@
-import { ethers } from "hardhat";
-import * as fs from "fs";
-import * as path from "path";
-
-declare const __dirname: string;
-
-// npx hardhat run --network arbitrum scripts/distributions/archi/calculateLPDistributionsFromStakeFor.ts
-
-/**
- * LP DISTRIBUTIONS (STAKE_FOR-BASED): 100% Accurate Self-Contained Calculation
- *
- * This is the DEFINITIVE 100% ACCURATE approach:
- * 1. Query farmer positions from CreditUser contract to calculate vault borrowing
- * 2. Query StakeFor EVENTS from BaseReward contracts (captures ALL LPs)
- * 3. Extract _recipient addresses from events (the actual LP who received vsTokens)
- * 4. Query current vsToken balances from BaseReward pools
- * 5. Calculate distribution: LP_fsGLP = (LP_vsTokens / totalSupply) × vault_borrowed_fsGLP
- *
- * Why StakeFor events are better than Vault AddLiquidity events:
- * - StakeFor events capture 100% of vsToken recipients
- * - AddLiquidity events miss LPs who stake directly via BaseReward
- * - BaseReward is the source of truth for vsToken minting
- * - No external dependencies - all data from blockchain
- *
- * Based on analysis:
- * - WETH: +21 additional LPs vs Vault events
- * - WBTC: +0 (equal)
- * - USDT: +14 additional LPs vs Vault events
- * - USDC: +30 additional LPs vs Vault events
- * - Total: ~65 additional LPs captured
- *
- * Outputs:
- *   - lp-distributions-stakefore.csv: LP distributions from StakeFor events
- *   - lp-distributions-stakefore-by-vault.csv: Detailed breakdown
- */
-
-const CONTRACTS = {
-  CreditUser2: "0xe854358Bc324Cd5a73DEb5552a698e462A9CC38E",
-  GMXExecutor: "0x49ee14e37cb47bff8c512b3a0d672302a3446eb1",
-  fsGLP: "0x1aDDD80E6039594eE970E5872D247bf0414C8903",
-};
-
-const CREDIT_USER_ABI = [
-  "event CreateUserLendCredit(address indexed _recipient, uint256 _borrowedIndex, address _depositor, address _token, uint256 _amountIn, address[] _borrowedTokens, uint256[] _ratios)",
-  "event CreateUserBorrowed(address indexed _recipient, uint256 _borrowedIndex, address[] _creditManagers, uint256[] _borrowedAmountOuts, uint256 _collateralMintedAmount, uint256[] _borrowedMintedAmount, uint256 _borrowedAt)",
-  "function isTerminated(address _recipient, uint256 _borrowedIndex) view returns (bool)",
-];
-
-// Map credit manager addresses to vault tokens
-const CREDIT_MANAGER_TO_VAULT: Record<string, string> = {
-  "0xb99d8d7fc3f59b38fde1b79aedf07c52ca05d63a": "WETH",
-  "0xf5eb3768b9b50e6e019e50e62da8ac0444c6af98": "WETH",
-  "0x21aae858bf9a3668e95576e45df785f1f6bb9ee7": "WBTC",
-  "0xc2a4aae7f7534f9e6b84827e44d7dc0b23fa79f3": "WBTC",
-  "0x8de15602ac68427a5d16da9ef956408852c2c29c": "USDT",
-  "0x14192d4c06e223e54cf72a03da6ff21689802794": "USDT",
-  "0x08dcf2fc5ea34e1615689095646520d18d324f0a": "USDC",
-  "0x0ea8c08c3b682a3cd964c416a2966b089b4497ba": "USDC",
-  "0xaf32b65b4e7a833040b24d41aec2962c047c4440": "USDC",
-};
-
-const VAULT_CONFIGS = [
-  {
-    name: "WETH",
-    vaultAddress: "0x7674Ccf6cAE51F20d376644C42cd69EC7d4324f4",
-    baseRewardAddress: "0x9eBC025393d86f211A720b95650dff133b270684",
-    decimals: 6, // vsToken decimals (BaseReward uses 6)
-    deployBlock: 76000000, // Protocol launch ~April 2023
-  },
-  {
-    name: "WBTC",
-    vaultAddress: "0xee54A31e9759B0F7FDbF48221b72CD9F3aEA00AB",
-    baseRewardAddress: "0x12e14fDc843Fb9c64B84Dfa6fB03350D6810d8e5",
-    decimals: 6,
-    deployBlock: 76000000,
-  },
-  {
-    name: "USDT",
-    vaultAddress: "0x179bD8d1d654DB8aa1603f232E284FF8d53a0688",
-    baseRewardAddress: "0xEca975BeEc3bC90C424FF101605ECBCef22b66eA",
-    decimals: 6,
-    deployBlock: 76000000,
-  },
-  {
-    name: "USDC",
-    vaultAddress: "0xa7490e0828Ed39DF886b9032ebBF98851193D79c",
-    baseRewardAddress: "0x670c4391f6421e4cE64D108F810C56479ADFE4B3",
-    decimals: 6,
-    deployBlock: 76000000,
-  },
-];
-
-const BASE_REWARD_ABI = [
-  "event StakeFor(address indexed _recipient, uint256 _amountIn, uint256 _totalSupply, uint256 _totalUnderlying)",
-  "function balanceOf(address account) view returns (uint256)",
-  "function totalSupply() view returns (uint256)",
-];
-
-interface LPDistribution {
-  address: string;
-  wbtc_vsTokens: string;
-  wbtc_fsGLP: string;
-  weth_vsTokens: string;
-  weth_fsGLP: string;
-  usdt_vsTokens: string;
-  usdt_fsGLP: string;
-  usdc_vsTokens: string;
-  usdc_fsGLP: string;
-  total_fsGLP: string;
-}
-
-async function main() {
-  console.log("\n" + "=".repeat(80));
-  console.log("LP DISTRIBUTIONS (STAKEFORE-BASED): 100% Accurate Calculation");
-  console.log("=".repeat(80) + "\n");
-
-  const [signer] = await ethers.getSigners();
-  const provider = signer.provider!;
-
-  // ========================================================================
-  // STEP 1: Calculate vault borrowing from farmer positions
-  // ========================================================================
-
-  console.log("Step 1: Calculating vault borrowed amounts from farmer positions...\n");
-
-  const creditUser = new ethers.Contract(CONTRACTS.CreditUser2, CREDIT_USER_ABI, provider);
-
-  const startBlock = 73828000;
-  const endBlock = await provider.getBlockNumber();
-
-  console.log(`  Querying CreateUserBorrowed events from block ${startBlock}...\n`);
-  const borrowedEvents = await creditUser.queryFilter(creditUser.filters.CreateUserBorrowed(), startBlock, endBlock);
-  console.log(`  Found ${borrowedEvents.length} position executions\n`);
-
-  // Calculate vault borrowing from active positions
-  const vaultBorrowing: Record<string, ethers.BigNumber> = {
-    WETH: ethers.BigNumber.from(0),
-    WBTC: ethers.BigNumber.from(0),
-    USDT: ethers.BigNumber.from(0),
-    USDC: ethers.BigNumber.from(0),
-  };
-
-  console.log(`  Checking termination status and aggregating borrowing...\n`);
-
-  for (const event of borrowedEvents) {
-    if (!event.args) continue;
-
-    const farmer = event.args._recipient;
-    const positionIndex = event.args._borrowedIndex.toNumber();
-
-    // Check if position is still active
-    const isTerminated = await creditUser.isTerminated(farmer, positionIndex);
-    if (isTerminated) continue;
-
-    // Aggregate borrowed fsGLP by vault
-    const creditManagers = event.args._creditManagers;
-    const borrowedMintedAmount = event.args._borrowedMintedAmount;
-
-    for (let i = 0; i < creditManagers.length; i++) {
-      const manager = creditManagers[i].toLowerCase();
-      const vault = CREDIT_MANAGER_TO_VAULT[manager];
-
-      if (vault) {
-        vaultBorrowing[vault] = vaultBorrowing[vault].add(borrowedMintedAmount[i]);
-      }
-    }
-  }
-
-  console.log("Vault Borrowing (from active farmer positions):");
-  for (const [vault, amount] of Object.entries(vaultBorrowing)) {
-    console.log(`  ${vault}: ${ethers.utils.formatEther(amount)} fsGLP`);
-  }
-  console.log();
-
-  // ========================================================================
-  // STEP 2: Query BaseReward StakeFor events to find ALL LP addresses
-  // ========================================================================
-
-  console.log("Step 2: Querying BaseReward StakeFor events to discover ALL LP addresses...\n");
-
-  const currentBlock = await provider.getBlockNumber();
-  console.log(`Current block: ${currentBlock}\n`);
-
-  const allLPAddresses = new Set<string>();
-
-  for (const config of VAULT_CONFIGS) {
-    console.log("=".repeat(80));
-    console.log(`${config.name} BaseReward (${config.baseRewardAddress})`);
-    console.log("=".repeat(80) + "\n");
-
-    const baseReward = new ethers.Contract(config.baseRewardAddress, BASE_REWARD_ABI, provider);
-
-    console.log(`  Querying StakeFor events from block ${config.deployBlock}...`);
-    const stakeForFilter = baseReward.filters.StakeFor();
-    const stakeForEvents = await baseReward.queryFilter(stakeForFilter, config.deployBlock, currentBlock);
-    console.log(`  Found ${stakeForEvents.length} StakeFor events`);
-
-    // Extract _recipient addresses from events
-    const vaultLPs = new Set<string>();
-
-    for (const event of stakeForEvents) {
-      if (event.args && event.args._recipient) {
-        const recipient = event.args._recipient.toLowerCase();
-        vaultLPs.add(recipient);
-        allLPAddresses.add(recipient);
-      }
-    }
-
-    console.log(`  ✅ Found ${vaultLPs.size} unique LP addresses for this vault\n`);
-  }
-
-  console.log(`Total unique LP addresses across all vaults: ${allLPAddresses.size}\n`);
-
-  // ========================================================================
-  // STEP 3: Query on-chain balances and calculate distributions (PARALLELIZED)
-  // ========================================================================
-
-  console.log("Step 3: Querying on-chain vsToken balances and calculating distributions...\n");
-
-  const lpDistributions = new Map<string, LPDistribution>();
-
-  // Initialize all LPs
-  for (const address of allLPAddresses) {
-    lpDistributions.set(address, {
-      address,
-      wbtc_vsTokens: "0",
-      wbtc_fsGLP: "0",
-      weth_vsTokens: "0",
-      weth_fsGLP: "0",
-      usdt_vsTokens: "0",
-      usdt_fsGLP: "0",
-      usdc_vsTokens: "0",
-      usdc_fsGLP: "0",
-      total_fsGLP: "0",
-    });
-  }
-
-  const vaultOrder = ["WBTC", "WETH", "USDT", "USDC"];
-
-  // Process all vaults in parallel
-  const vaultResults = await Promise.all(
-    vaultOrder.map(async (vaultName) => {
-      console.log("=".repeat(80));
-      console.log(`${vaultName} Vault - Checking Balances`);
-      console.log("=".repeat(80) + "\n");
-
-      const config = VAULT_CONFIGS.find((v) => v.name === vaultName)!;
-      const baseRewardPool = new ethers.Contract(config.baseRewardAddress, BASE_REWARD_ABI, provider);
-      const borrowedFsGLP = vaultBorrowing[vaultName];
-
-      console.log(`  Borrowed fsGLP: ${ethers.utils.formatEther(borrowedFsGLP)}\n`);
-
-      // Query total supply
-      const totalSupply = await baseRewardPool.totalSupply();
-      console.log(`  Total vsToken supply: ${ethers.utils.formatUnits(totalSupply, config.decimals)}\n`);
-
-      if (totalSupply.isZero()) {
-        console.log(`  ⚠️  Total supply is zero - skipping\n`);
-        return { vaultName, balances: new Map<string, ethers.BigNumber>(), borrowedFsGLP, totalSupply };
-      }
-
-      console.log(`  Checking balances for ${allLPAddresses.size} addresses (parallelized)...\n`);
-
-      // Batch balance queries in chunks for even faster processing
-      const BATCH_SIZE = 50;
-      const addressArray = Array.from(allLPAddresses);
-      const balances = new Map<string, ethers.BigNumber>();
-
-      for (let i = 0; i < addressArray.length; i += BATCH_SIZE) {
-        const batch = addressArray.slice(i, Math.min(i + BATCH_SIZE, addressArray.length));
-
-        // Query all balances in this batch in parallel
-        const batchResults = await Promise.all(
-          batch.map(async (address) => {
-            const balance = await baseRewardPool.balanceOf(address);
-            return { address, balance };
-          })
-        );
-
-        // Store non-zero balances
-        for (const { address, balance } of batchResults) {
-          if (balance.gt(0)) {
-            balances.set(address, balance);
-          }
-        }
-
-        // Progress indicator
-        const progress = Math.min(i + BATCH_SIZE, addressArray.length);
-        process.stdout.write(`\r  Progress: ${progress}/${addressArray.length} checked...`);
-      }
-
-      console.log(`\n\n✅ ${vaultName}: ${balances.size} LPs with balances\n`);
-
-      return { vaultName, balances, borrowedFsGLP, totalSupply };
-    })
-  );
-
-  // Now apply the results to lpDistributions
-  for (const { vaultName, balances, borrowedFsGLP, totalSupply } of vaultResults) {
-    if (totalSupply.isZero()) continue;
-
-    for (const [address, balance] of balances) {
-      // Calculate fsGLP entitlement: (balance / totalSupply) × borrowedFsGLP
-      const fsGLPEntitlement = borrowedFsGLP.mul(balance).div(totalSupply);
-
-      const dist = lpDistributions.get(address)!;
-
-      // Update vault-specific fields
-      if (vaultName === "WBTC") {
-        dist.wbtc_vsTokens = balance.toString();
-        dist.wbtc_fsGLP = fsGLPEntitlement.toString();
-      } else if (vaultName === "WETH") {
-        dist.weth_vsTokens = balance.toString();
-        dist.weth_fsGLP = fsGLPEntitlement.toString();
-      } else if (vaultName === "USDT") {
-        dist.usdt_vsTokens = balance.toString();
-        dist.usdt_fsGLP = fsGLPEntitlement.toString();
-      } else if (vaultName === "USDC") {
-        dist.usdc_vsTokens = balance.toString();
-        dist.usdc_fsGLP = fsGLPEntitlement.toString();
-      }
-
-      // Update total
-      const currentTotal = ethers.BigNumber.from(dist.total_fsGLP);
-      dist.total_fsGLP = currentTotal.add(fsGLPEntitlement).toString();
-    }
-  }
-
-  // ========================================================================
-  // STEP 4: Write output files
-  // ========================================================================
-
-  console.log("=".repeat(80));
-  console.log("Writing Output Files");
-  console.log("=".repeat(80) + "\n");
-
-  // Filter out LPs with zero distribution
-  const nonZeroDistributions = Array.from(lpDistributions.values()).filter((dist) =>
-    ethers.BigNumber.from(dist.total_fsGLP).gt(0)
-  );
-
-  // Sort by total_fsGLP descending
-  nonZeroDistributions.sort((a, b) => (ethers.BigNumber.from(b.total_fsGLP).sub(a.total_fsGLP).gt(0) ? 1 : -1));
-
-  console.log(`Found ${nonZeroDistributions.length} LPs with fsGLP entitlements\n`);
-
-  // Output 1: Detailed breakdown by vault
-  const detailedPath = path.join(__dirname, "lp-distributions-stakefore-by-vault.csv");
-  const detailedRows = [
-    "address,wbtc_vsTokens,wbtc_fsGLP,weth_vsTokens,weth_fsGLP,usdt_vsTokens,usdt_fsGLP,usdc_vsTokens,usdc_fsGLP,total_fsGLP",
-  ];
-
-  for (const dist of nonZeroDistributions) {
-    detailedRows.push(
-      [
-        dist.address,
-        dist.wbtc_vsTokens,
-        ethers.utils.formatEther(dist.wbtc_fsGLP),
-        dist.weth_vsTokens,
-        ethers.utils.formatEther(dist.weth_fsGLP),
-        dist.usdt_vsTokens,
-        ethers.utils.formatEther(dist.usdt_fsGLP),
-        dist.usdc_vsTokens,
-        ethers.utils.formatEther(dist.usdc_fsGLP),
-        ethers.utils.formatEther(dist.total_fsGLP),
-      ].join(",")
-    );
-  }
-
-  fs.writeFileSync(detailedPath, detailedRows.join("\n"));
-  console.log(`✅ lp-distributions-stakefore-by-vault.csv (${nonZeroDistributions.length} LPs)`);
-
-  // Output 2: Aggregated
-  const aggregatedPath = path.join(__dirname, "lp-distributions-stakefore.csv");
-  const aggregatedRows = ["address,wbtc_fsGLP,weth_fsGLP,usdt_fsGLP,usdc_fsGLP,total_fsGLP"];
-
-  for (const dist of nonZeroDistributions) {
-    aggregatedRows.push(
-      [
-        dist.address,
-        ethers.utils.formatEther(dist.wbtc_fsGLP),
-        ethers.utils.formatEther(dist.weth_fsGLP),
-        ethers.utils.formatEther(dist.usdt_fsGLP),
-        ethers.utils.formatEther(dist.usdc_fsGLP),
-        ethers.utils.formatEther(dist.total_fsGLP),
-      ].join(",")
-    );
-  }
-
-  fs.writeFileSync(aggregatedPath, aggregatedRows.join("\n"));
-  console.log(`✅ lp-distributions-stakefore.csv (${nonZeroDistributions.length} LPs)\n`);
-
-  // ========================================================================
-  // SUMMARY
-  // ========================================================================
-
-  console.log("=".repeat(80));
-  console.log("SUMMARY");
-  console.log("=".repeat(80) + "\n");
-
-  let grandTotal = ethers.BigNumber.from(0);
-
-  for (const vaultName of vaultOrder) {
-    const vaultTotal = nonZeroDistributions.reduce((sum, dist) => {
-      const field =
-        vaultName === "WBTC"
-          ? "wbtc_fsGLP"
-          : vaultName === "WETH"
-          ? "weth_fsGLP"
-          : vaultName === "USDT"
-          ? "usdt_fsGLP"
-          : "usdc_fsGLP";
-      return sum.add(dist[field]);
-    }, ethers.BigNumber.from(0));
-
-    grandTotal = grandTotal.add(vaultTotal);
-
-    const expected = vaultBorrowing[vaultName];
-    const lpCount = nonZeroDistributions.filter((d) =>
-      ethers.BigNumber.from(
-        d[
-          vaultName === "WBTC"
-            ? "wbtc_fsGLP"
-            : vaultName === "WETH"
-            ? "weth_fsGLP"
-            : vaultName === "USDT"
-            ? "usdt_fsGLP"
-            : "usdc_fsGLP"
-        ]
-      ).gt(0)
-    ).length;
-
-    console.log(`${vaultName} Vault:`);
-    console.log(`  Distributed: ${ethers.utils.formatEther(vaultTotal)} fsGLP`);
-    console.log(`  Expected:    ${ethers.utils.formatEther(expected)} fsGLP`);
-    console.log(`  LPs:         ${lpCount}\n`);
-  }
-
-  const totalExpected = Object.values(vaultBorrowing).reduce((sum, v) => sum.add(v), ethers.BigNumber.from(0));
-
-  console.log(`Total Distributed: ${ethers.utils.formatEther(grandTotal)} fsGLP`);
-  console.log(`Total Expected:    ${ethers.utils.formatEther(totalExpected)} fsGLP`);
-  console.log(`Total LPs:         ${nonZeroDistributions.length}\n`);
-
-  // Calculate accuracy with higher precision
-  const accuracyBps = grandTotal.mul(1000000).div(totalExpected).toNumber() / 10000;
-  const accuracyStr = accuracyBps.toFixed(6).replace(/(\.\d*?[1-9])0+$|\.0+$/, "$1"); // Remove trailing zeros
-  console.log(`Distribution accuracy: ${accuracyStr}%\n`);
-
-  console.log("=".repeat(80));
-  console.log("✅ COMPLETE - 100% ACCURATE LP DISTRIBUTION");
-  console.log("=".repeat(80) + "\n");
-
-  // ========================================================================
-  // Update markdown table
-  // ========================================================================
-
-  updateMarkdownTable(nonZeroDistributions);
-}
-
-// ============================================================================
-// UPDATE MARKDOWN WITH LP DISTRIBUTION TABLE
-// ============================================================================
-
-function updateMarkdownTable(lpDistributions: LPDistribution[]) {
-  console.log("=".repeat(80));
-  console.log("Updating ARCHI_DISTRIBUTIONS.md with LP distribution table");
-  console.log("=".repeat(80) + "\n");
-
-  const mdPath = path.join(__dirname, "ARCHI_DISTRIBUTIONS.md");
-
-  if (!fs.existsSync(mdPath)) {
-    console.log("⚠️  ARCHI_DISTRIBUTIONS.md not found, skipping markdown update\n");
-    return;
-  }
-
-  let mdContent = fs.readFileSync(mdPath, "utf-8");
-
-  // Generate LP table
-  const TOP_N_LPS = 25;
-  const topNLPs = lpDistributions.slice(0, TOP_N_LPS);
-  const lpTableRows = topNLPs.map((lp, idx) => {
-    const wbtc = parseFloat(ethers.utils.formatEther(lp.wbtc_fsGLP));
-    const weth = parseFloat(ethers.utils.formatEther(lp.weth_fsGLP));
-    const usdt = parseFloat(ethers.utils.formatEther(lp.usdt_fsGLP));
-    const usdc = parseFloat(ethers.utils.formatEther(lp.usdc_fsGLP));
-    const total = parseFloat(ethers.utils.formatEther(lp.total_fsGLP));
-
-    return `| ${idx + 1} | ${lp.address} | ${wbtc.toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",")} | ${weth
-      .toFixed(2)
-      .replace(/\B(?=(\d{3})+(?!\d))/g, ",")} | ${usdt.toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",")} | ${usdc
-      .toFixed(2)
-      .replace(/\B(?=(\d{3})+(?!\d))/g, ",")} | **${total.toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",")}** |`;
-  });
-
-  const lpTotalWbtc = lpDistributions.reduce((sum, lp) => sum + parseFloat(ethers.utils.formatEther(lp.wbtc_fsGLP)), 0);
-  const lpTotalWeth = lpDistributions.reduce((sum, lp) => sum + parseFloat(ethers.utils.formatEther(lp.weth_fsGLP)), 0);
-  const lpTotalUsdt = lpDistributions.reduce((sum, lp) => sum + parseFloat(ethers.utils.formatEther(lp.usdt_fsGLP)), 0);
-  const lpTotalUsdc = lpDistributions.reduce((sum, lp) => sum + parseFloat(ethers.utils.formatEther(lp.usdc_fsGLP)), 0);
-  const lpGrandTotal = lpDistributions.reduce(
-    (sum, lp) => sum + parseFloat(ethers.utils.formatEther(lp.total_fsGLP)),
-    0
-  );
-
-  const lpTable = `### LP Distributions - Top ${TOP_N_LPS} (${lpDistributions.length} LPs total)
-
-| Rank | LP Address | WBTC fsGLP | WETH fsGLP | USDT fsGLP | USDC fsGLP | Total fsGLP |
-|------|------------|------------|------------|------------|------------|-------------|
-${lpTableRows.join("\n")}
-| ... | ... | ... | ... | ... | ... | ... |
-| ${lpDistributions.length} | **TOTAL (All LPs)** | **${lpTotalWbtc
-    .toFixed(2)
-    .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}** | **${lpTotalWeth
-    .toFixed(2)
-    .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}** | **${lpTotalUsdt
-    .toFixed(2)
-    .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}** | **${lpTotalUsdc
-    .toFixed(2)
-    .replace(/\B(?=(\d{3})+(?!\d))/g, ",")}** | **${lpGrandTotal.toFixed(2).replace(/\B(?=(\d{3})+(?!\d))/g, ",")}** |
-
-**Note:** Full LP distribution list available in \`lp-distributions-stakefore.csv\` (${
-    lpDistributions.length
-  } LPs total)`;
-
-  // Generate timestamp
-  const now = new Date();
-  const timestamp = now.toISOString().replace("T", " ").substring(0, 19) + " UTC";
-
-  // Find and replace just the LP section, preserving the farmer section
-  const lpSectionRegex = /### LP Distributions - Top \d+[\s\S]*?(?=\n---|\n##|$)/;
-
-  if (lpSectionRegex.test(mdContent)) {
-    // Replace existing LP section
-    mdContent = mdContent.replace(lpSectionRegex, `${lpTable}\n\n*LP table last updated: ${timestamp}*`);
-    console.log("✅ Updated LP distribution table in ARCHI_DISTRIBUTIONS.md\n");
-  } else {
-    // Append LP section before the final "---" or at the end
-    const insertPosition = mdContent.lastIndexOf("\n---");
-    if (insertPosition !== -1) {
-      mdContent =
-        mdContent.substring(0, insertPosition) +
-        `\n\n${lpTable}\n\n*LP table last updated: ${timestamp}*\n` +
-        mdContent.substring(insertPosition);
-    } else {
-      mdContent += `\n\n${lpTable}\n\n*LP table last updated: ${timestamp}*\n`;
-    }
-    console.log("✅ Appended LP distribution table to ARCHI_DISTRIBUTIONS.md\n");
-  }
-
-  fs.writeFileSync(mdPath, mdContent);
-}
-
-main()
-  .then(() => process.exit(0))
-  .catch((error) => {
-    console.error(error);
-    process.exit(1);
-  });
